Рефакторинг по замечаниям
