--- src/upload-form.js
@@ -1,94 +1,0 @@
-'use strict';
-
-/** @enum {string} */
-var FileType = {
-  'GIF': '',
-  'JPEG': '',
-  'PNG': '',
-  'SVG+XML': ''
-};
-
-/** @enum {number} */
-var Action = {
-  ERROR: 0,
-  UPLOADING: 1,
-  CUSTOM: 2
-};
-
-/**
- * Регулярное выражение, проверяющее тип загружаемого файла. Составляется
- * из ключей FileType.
- * @type {RegExp}
- */
-var fileRegExp = new RegExp('^image/(' + Object.keys(FileType).join('|').replace('\+', '\\+') + ')$', 'i');
-
-/**
- * Форма загрузки изображения.
- * @type {HTMLFormElement}
- */
-var uploadForm = document.forms['upload-select-image'];
-
-/**
- * @type {HTMLElement}
- */
-var uploadMessage = document.querySelector('.upload-message');
-
-/**
- * @type {HTMLElement}
- */
-var backgroundElement = document.querySelector('.upload');
-
-/**
- * @param {Action} action
- * @param {string=} message
- * @return {Element}
- */
-function showMessage(action, message) {
-  var isError = false;
-
-  switch (action) {
-    case Action.UPLOADING:
-      message = message || 'Кексограмим&hellip;';
-      break;
-
-    case Action.ERROR:
-      isError = true;
-      message = message || 'Неподдерживаемый формат файла<br> <a href="' + document.location + '">Попробовать еще раз</a>.';
-      break;
-  }
-
-  uploadMessage.querySelector('.upload-message-container').innerHTML = message;
-  uploadMessage.classList.remove('invisible');
-  uploadMessage.classList.toggle('upload-message-error', isError);
-  return uploadMessage;
-}
-
-function hideMessage() {
-  uploadMessage.classList.add('invisible');
-}
-
-/**
- * Ставит одну из трех случайных картинок на фон формы загрузки.
- */
-function updateBackground() {
-  var images = [
-    'img/logo-background-1.jpg',
-    'img/logo-background-2.jpg',
-    'img/logo-background-3.jpg'
-  ];
-
-  var randomImageNumber = Math.round(Math.random() * (images.length - 1));
-  backgroundElement.style.backgroundImage = 'url(' + images[randomImageNumber] + ')';
-}
-
-updateBackground();
-
-module.exports = {
-  fileRegExp: fileRegExp,
-  uploadForm: uploadForm,
-  showMessage: showMessage,
-  UPLOADING: Action.UPLOADING,
-  ERROR: Action.ERROR,
-  hideMessage: hideMessage,
-  updateBackground: updateBackground
-};
