+++ src/utilities.js
@@ -1,0 +1,23 @@
+'use strict';
+
+/** @enum {string} */
+var FileType = {
+  'GIF': '',
+  'JPEG': '',
+  'PNG': '',
+  'SVG+XML': ''
+};
+/**
+ * Регулярное выражение, проверяющее тип загружаемого файла. Составляется
+ * из ключей FileType.
+ * @type {RegExp}
+ */
+var fileRegExp = new RegExp('^image/(' + Object.keys(FileType).join('|').replace('\+', '\\+') + ')$', 'i');
+
+function isImgFile(type) {
+  return fileRegExp.test(type);
+}
+
+module.exports = {
+  isImgFile: isImgFile
+};
